FloodSight/streamlit_app/requirements.txt
streamlit>=1.30.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
tensorflow>=2.14.0
joblib>=1.3.0
scikit-learn>=1.3.0
seaborn>=0.13.2
pillow>=10.1.0
python-dotenv>=1.1.1
